requirements.txt

numpy
openai
google-genai
python-dotenv
